# Optimize tridiagonal matrix solver