Had some code laying around that I needed for verifying a GPU implementation, and it seems to outperform our current generic fallback:

```julia
julia> function benchmark(;N=256, T=Float32)
           t = Tridiagonal(rand(T, N, N))
           rhs = rand(T, N)

           suite = BenchmarkGroup()
           suite["Base"] = @benchmarkable invoke(\, Tuple{AbstractMatrix, AbstractVector}, $t, $rhs)
           suite["new"] = @benchmarkable $t \ $rhs

           warmup(suite)
           results = run(suite)
           judge(median(results["new"]), median(results["Base"]))
       end

julia> benchmark(N=4)
  "Base" => Trial(219.000 ns)
  "new" => Trial(91.000 ns)

  time:   -58.44% => improvement (5.00% tolerance)
  memory: -73.33% => improvement (1.00% tolerance)


julia> benchmark(N=64)
  "Base" => Trial(1.186 μs)
  "new" => Trial(661.000 ns)

  time:   -49.93% => improvement (5.00% tolerance)
  memory: -72.37% => improvement (1.00% tolerance)


julia> benchmark(N=1024)
  "Base" => Trial(18.241 μs)
  "new" => Trial(10.111 μs)

  time:   -47.03% => improvement (5.00% tolerance)
  memory: -71.43% => improvement (1.00% tolerance)
```

I'm not well versed in linalg so probably making some silly mistakes here.