@andreasnoack 2019-08-02
I suspect at least part of the reason for the time difference is the lack of pivoting in the version here. It makes it unstable for some matrices. However, the current implementation without pivoting isn't faster than the one with pivoting. I'll have to take a closer look.
```julia
julia> function benchmark(;N=256, T=Float32)
                  t = Tridiagonal(rand(T, N, N))
                  rhs = rand(T, N)

                  suite = BenchmarkGroup()
                  suite["Base"] = @benchmarkable invoke(\, Tuple{AbstractMatrix, AbstractVector}, $t, $rhs)
                  suite["BaseNoPivot"] = @benchmarkable lu($t, Val(false)) \ $rhs
                  suite["new"] = @benchmarkable $t \ $rhs

                  warmup(suite)
                  results = run(suite)
                  display(judge(median(results["new"]), median(results["Base"])))
                  display(judge(median(results["Base"]), median(results["BaseNoPivot"])))
              end
benchmark (generic function with 1 method)

julia> benchmark(N=64)
(1/3) benchmarking "Base"...
done (took 0.062119647 seconds)
(2/3) benchmarking "new"...
done (took 0.060346363 seconds)
(3/3) benchmarking "BaseNoPivot"...
done (took 0.070538019 seconds)
BenchmarkTools.TrialJudgement:
  time:   -47.63% => improvement (5.00% tolerance)
  memory: -72.37% => improvement (1.00% tolerance)

BenchmarkTools.TrialJudgement:
  time:   -6.24% => improvement (5.00% tolerance)
  memory: +0.00% => invariant (1.00% tolerance)
```